+++ R/plots.R
@@ -242,6 +242,8 @@ roboplotr_dependencies <- function(p,
 #' @param modebar Function. Use [set_modebar()].
 #' @param info_text Character. Adds an info button to the modebar with this text, along with plot title and caption.
 #' @param updatemenu Function. Use [set_updatemenu()] for detailed control.
+#' @param roboplot_options Character. A name of roboplot options set with [set_roboplot_options(name)]. 
+#' See documentation of [set_roboplot_options()] for details.
 #' @param ... Placeholder for other parameters.
 #' @returns A list of classes "htmlwidget", "plotly", and  "roboplotr.roboplot"
 #' @examples
@@ -251,7 +253,7 @@ roboplotr_dependencies <- function(p,
 #'
 #'
 #' d <- energiantuonti |>
-#'   dplyr::filter(Alue %in% c("Kanada","Norja","Yhdistynyt kuningaskunta"))
+#'   dplyr::filter(Alue %in% c("USA","Norja","Iso-Britannia"))
 #' d1 <- d |> dplyr::filter(Suunta == "Tuonti")
 #' d1 |> roboplot(color = Alue,
 #'                title = "Energian tuonti",
@@ -263,7 +265,7 @@ roboplotr_dependencies <- function(p,
 #' # works). Caption may be further specified with the helper function
 #' # `set_caption()` (see documentation for more control).
 #' d1 |>
-#'   dplyr::filter(Alue == "Yhdistynyt kuningaskunta") |>
+#'   dplyr::filter(Alue == "Iso-Britannia") |>
 #'   roboplot(Alue,"Energian tuonti Yhdistyneest\uE4 kuningaskunnasta","Milj. \u20AC",
 #'            caption = set_caption(text = "Tilastokeskus")
 #'   )
@@ -285,6 +287,10 @@ roboplotr_dependencies <- function(p,
 #' # Reset to defaults
 #' set_roboplot_options(reset = TRUE)
 #'
+#' # You can omit caption by setting it to NA.
+#'
+#' d1 |> roboplot(Alue, "Energian tuonti","Milj. \u20AC",caption = NA)
+#' 
 #' # Legend can also be omitted by giving `legend_position` of NA. Height and
 #' # width can also be specified, while for most uses width specification is
 #' # unnecessary, as roboplotr is designed for plots with responsive widths.
@@ -324,9 +330,9 @@ roboplotr_dependencies <- function(p,
 #' # trace by giving named character vectors as the appropriate arguments.
 #' # Barmode or scatter type is controlled by `plot_mode`.
 #' d1 |> roboplot(Alue,"Energian tuonti ja vienti","Milj. \u20AC","Tilastokeskus",
-#'              trace_color =  c("Kanada" = "red","Norja" = "blue", .other = "black"),
-#'              plot_type = c("Norja" = "scatter","Kanada" = "bar",".other" = "scatter"),
-#'              plot_mode = c("Yhdistynyt kuningaskunta" = "scatter",
+#'              trace_color =  c("USA" = "red","Norja" = "blue", .other = "black"),
+#'              plot_type = c("Norja" = "scatter","USA" = "bar",".other" = "scatter"),
+#'              plot_mode = c("Iso-Britannia" = "scatter",
 #'                            "Norja" = "scatter+line", ".bar" = "dodge",
 #'                            ".scatter" = "line"
 #'              ))
@@ -334,7 +340,7 @@ roboplotr_dependencies <- function(p,
 #' # If you omit ".other" from the trace colors,` roboplot()` will give the rest of
 #' # of the traces colors from the default colors set in `set_roboplot_options()`
 #' d1 |> roboplot(Alue,"Energian tuonti ja vienti","Milj. \u20AC","Tilastokeskus",
-#'                trace_color = c("Kanada" = "pink"))
+#'                trace_color = c("USA" = "pink"))
 #'
 #' # But if you need more control, you're better off just excplicity specifying
 #' # the colors.
@@ -425,23 +431,23 @@ roboplotr_dependencies <- function(p,
 #' # You can use `secondary_yaxis` to define which observations from 'color' use
 #' # go to a secondary yaxis on the right.
 #' d2 |>
-#'   dplyr::filter(Suunta == "Tuonti", Alue %in% c("Yhdistynyt kuningaskunta", "Kanada", "Norja")) |>
+#'   dplyr::filter(Suunta == "Tuonti", Alue %in% c("Iso-Britannia", "USA", "Norja")) |>
 #'   roboplot(Alue,
 #'            "Energian tuonti",
 #'            "Milj. \u20AC",
 #'            "Tilastokeskus",
-#'            secondary_yaxis = "Yhdistynyt kuningaskunta",
+#'            secondary_yaxis = "Iso-Britannia",
 #'            zeroline = 1000)
 #'
 #' # Furthermore, you can use `set_axes()` in `plot_axes` for further control, like
 #' # titles. Documentation for `set_axes()` has more detailed examples.
 #' d2 |>
-#'   dplyr::filter(Suunta == "Tuonti", Alue %in% c("Yhdistynyt kuningaskunta", "Kanada", "Norja")) |>
+#'   dplyr::filter(Suunta == "Tuonti", Alue %in% c("Iso-Britannia", "USA", "Norja")) |>
 #'   roboplot(Alue,
 #'            "Energian tuonti",
 #'            "Milj. \u20AC",
 #'            "Tilastokeskus",
-#'            plot_axes = set_axes(y2 = "Yhdistynyt kuningaskunta"))
+#'            plot_axes = set_axes(y2 = "Iso-Britannia"))
 #'
 #' # Pie plots are possible too, but `pattern` is currently ignored by `plotly`
 #' # and thus by `roboplot()`.
@@ -585,10 +591,17 @@ roboplot <- function(d = NULL,
                      artefacts = getOption("roboplot.artefacts")$auto,
                      info_text = NULL,
                      updatemenu = NULL,
+                     roboplot_options = NULL,
                      legend_position = deprecated(),
                      legend_maxwidth = deprecated(),
                      ...) {
+  
+  .reset_options <- list(roboplot_options = NULL, reset = F)
+  on.exit(roboplotr_reset_temp_options(.reset_options))
+  
   margin <- NA # will this be used at all? Probably not.
+  
+  .reset_options <- roboplotr_temp_options(roboplot_options)
 
   roboplotr_typecheck(d, "data.frame", NULL, allow_null = F, allow_na = T)
 
@@ -770,7 +783,7 @@ roboplot <- function(d = NULL,
   }
 
 
-  plot_axes <- roboplotr_expand_axis_limits(plot_axes, d)
+  plot_axes <- roboplotr_expand_axis_limits(plot_axes, d, zeroline)
 
   if (xaxis_ceiling != "default" &
       all(is.na(plot_axes$xlim)) &