+++ R/robotable.R
@@ -385,7 +385,7 @@ roboplotr_set_robotable_fonts <-
 #' # handles the formatting automatically.
 #' #
 #' d <- energiantuonti |>
-#'   dplyr::filter(Alue %in% c("Ruotsi","Kanada")) |>
+#'   dplyr::filter(Alue %in% c("Ruotsi","USA")) |>
 #'   tidyr::unite(Tiedot, Alue, Suunta, sep = ", ") |>
 #'   dplyr::arrange(Tiedot, time) |>
 #'   tidyr::pivot_wider(names_from = Tiedot) |>
@@ -415,7 +415,7 @@ roboplotr_set_robotable_fonts <-
 #'   class = c("compact", "nowrap"),
 #'   searchable = FALSE,
 #'   sortable = FALSE,
-#'   col_widths = c("Kanada, Vienti" = 50)
+#'   col_widths = c("USA, Vienti" = 50)
 #' )
 robotable <-
   function(d,