+++ tests/testthat/helper-plot.R
@@ -1,5 +1,5 @@
 attr(energiantuonti, "frequency") <- "Quarterly"
 roboplot_tester <- energiantuonti |>
-  dplyr::filter(Alue %in% c("Kanada","Norja","Yhdistynyt kuningaskunta"),
+  dplyr::filter(Alue %in% c("USA","Norja","Iso-Britannia"),
                 Suunta == "Tuonti") |>
   plotly::plot_ly(x = ~ time,y = ~ value)